+++ src/caselawclient/models/press_summaries.py
@@ -27,3 +27,7 @@ class PressSummary(NeutralCitationMixin, Document):
                 "akn": "http://docs.oasis-open.org/legaldocml/ns/akn/3.0",
             },
         )
+
+    @property
+    def best_human_identifier(self) -> str:
+        return self.neutral_citation